Create Library Launcher abstraction to wrap calls to registerForActivityResult within library code.
